# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for TechGuild Platform.",
        "title": "TechGuild Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "TechGuild Team",
            "email": "support@techguild.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "https://techguild-backend.onrender.com",
    "basePath": "/",
    "paths": {
...
        "/settings/account": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the authenticated user's account settings.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Update account settings",
                "parameters": [
                    {
                        "description": "Account settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UpdateAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/settings/notifications": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the authenticated user's notification preferences.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Update notification settings",
                "parameters": [
                    {
                        "description": "Notification settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateNotificationsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UpdateNotificationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/settings/privacy": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the authenticated user's privacy preferences.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Update privacy settings",
                "parameters": [
                    {
                        "description": "Privacy settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePrivacyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UpdatePrivacySettingsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.UpdateAccountRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                },
                "password": {
                    "description": "For verification if changing email/password",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateNotificationsRequest": {
            "type": "object",
            "required": [
                "preferences"
            ],
            "properties": {
                "preferences": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "dto.UpdatePrivacyRequest": {
            "type": "object",
            "required": [
                "profile_visibility"
            ],
            "properties": {
                "profile_visibility": {
                    "type": "string",
                    "enum": [
                        "public",
                        "private"
                    ]
                }
            }
        },
...
        "swagger.UpdateAccountResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "account settings updated successfully"
                }
            }
        },
        "swagger.UpdateNotificationsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "notifications updated successfully"
                }
            }
        },
        "swagger.UpdatePrivacySettingsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "privacy settings updated successfully"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}